Move metric name check to the registry
