+++ js/modules/k6/metrics/metrics.go
@@ -22,8 +22,6 @@ package metrics
 
 import (
 	"errors"
-	"fmt"
-	"regexp"
 	"time"
 
 	"github.com/dop251/goja"
@@ -32,14 +30,6 @@ import (
 	"go.k6.io/k6/js/modules"
 	"go.k6.io/k6/stats"
 )
-
-var nameRegexString = "^[\\p{L}\\p{N}\\._ !\\?/&#\\(\\)<>%-]{1,128}$"
-
-var compileNameRegex = regexp.MustCompile(nameRegexString)
-
-func checkName(name string) bool {
-	return compileNameRegex.Match([]byte(name))
-}
 
 type Metric struct {
 	metric *stats.Metric
@@ -56,11 +46,6 @@ func (mi *ModuleInstance) newMetric(call goja.ConstructorCall, t stats.MetricTyp
 	}
 	rt := mi.GetRuntime()
 	c, _ := goja.AssertFunction(rt.ToValue(func(name string, isTime ...bool) (*goja.Object, error) {
-		// TODO: move verification outside the JS
-		if !checkName(name) {
-			return nil, common.NewInitContextError(fmt.Sprintf("Invalid metric name: '%s'", name))
-		}
-
 		valueType := stats.Default
 		if len(isTime) > 0 && isTime[0] {
 			valueType = stats.Time
