+++ js/modules/k6/metrics/metrics_test.go
@@ -155,28 +155,6 @@ func TestMetrics(t *testing.T) {
 	}
 }
 
-func TestMetricNames(t *testing.T) {
-	t.Parallel()
-	testMap := map[string]bool{
-		"simple":       true,
-		"still_simple": true,
-		"":             false,
-		"@":            false,
-		"a":            true,
-		"special\n\t":  false,
-		// this has both hangul and japanese numerals .
-		"hello.World_in_한글一안녕一세상": true,
-		// too long
-		"tooolooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooog": false,
-	}
-
-	for key, value := range testMap {
-		t.Run(key, func(t *testing.T) {
-			assert.Equal(t, value, checkName(key), key)
-		})
-	}
-}
-
 func TestMetricGetName(t *testing.T) {
 	t.Parallel()
 	rt := goja.New()
